+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file, as of versi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.3] - 2021-07-14
+- add test case for deploy contract using blob hex
+    * `yarn test:unit src/providers/jsonrpc-provider.spec.ts --testNamePattern="deploy contract"`
+
+- add test case for decode SignedMessage
+    * `yarn test:unit src/utils/signed-message.spec.ts --testNamePattern="decode"`
+
+- add chain_id in SignedMessage
+
+- add encoding.packageHexToTransactionPayload / packageHexToTransactionPayloadHex
+
+- support contract.dry_run_raw
+
 ## [1.4.2] - 2021-07-11
 - add 2 functions in utils.signedMessage
     * encodeSignedMessage